unit tests reorganized
