+++ test/Selenium.WebDriver.Extensions.Core.Tests/Loaders/QuerySelectorLoaderTests.cs
@@ -1,6 +1,7 @@
 ﻿namespace Selenium.WebDriver.Extensions.Core.Tests
 {
     using NUnit.Framework;
+
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
