+++ test/Selenium.WebDriver.Extensions.Core.Tests/Runners/JavaScriptRunnerTests.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using NUnit.Framework;
+
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
