+++ test/Selenium.WebDriver.Extensions.Core.Tests/Selectors/QuerySelectorTests.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Collections;
     using NUnit.Framework;
+
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
