+++ test/Selenium.WebDriver.Extensions.Core.Tests/Selectors/TagNameSelectorTests.cs
@@ -3,6 +3,7 @@
     using System.Collections;
     using NUnit.Framework;
     using By = Selenium.WebDriver.Extensions.Core.By;
+
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
