+++ test/Selenium.WebDriver.Extensions.IntegrationTests/WebDriverExtensions.JQuerySelectorTests.cs
@@ -7,6 +7,7 @@
     using Selenium.WebDriver.Extensions.IntegrationTests.Utils;
     using Selenium.WebDriver.Extensions.JQuery;
     using By = Selenium.WebDriver.Extensions.By;
+
     [TestFixture(
         WebBrowser.PhantomJs,
         "https://cdn.rawgit.com/RaYell/selenium-webdriver-extensions/63d90edf0560889e753b5fa464f7bf825ee39168/test/Selenium.WebDriver.Extensions.IntegrationTests/TestCases/JQuery/Loaded.html")]
