+++ test/Selenium.WebDriver.Extensions.JQuery.Tests/Extensions/WebDriverExtensionsTests.cs
@@ -1,4 +1,4 @@
-﻿namespace Selenium.WebDriver.Extensions.Tests
+﻿namespace Selenium.WebDriver.Extensions.JQuery.Tests
 {
     using System;
     using System.Collections;
@@ -10,8 +10,8 @@
     using OpenQA.Selenium;
     using Selenium.WebDriver.Extensions.Core;
     using Selenium.WebDriver.Extensions.JQuery;
-    using Selenium.WebDriver.Extensions.Sizzle;
-    using By = Selenium.WebDriver.Extensions.By;
+    using By = Selenium.WebDriver.Extensions.JQuery.By;
+    
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
@@ -43,30 +43,6 @@
             }
         }
 
-        private static IEnumerable LoadSizzleTestCases
-        {
-            get
-            {
-                yield return new TestCaseData("master", null, new object[] { true });
-                yield return new TestCaseData("1.11.1", null, new object[] { false, true, true });
-                yield return new TestCaseData("master", TimeSpan.FromSeconds(1), new object[] { false })
-                    .Throws(typeof(WebDriverTimeoutException));
-            }
-        }
-
-        private static IEnumerable LoadSizzleWithUriTestCases
-        {
-            get
-            {
-                yield return new TestCaseData(new Uri("http://my.com/sizzle.js"), null, new object[] { true });
-                yield return new TestCaseData(null, null, new object[] { false, true, true });
-                yield return new TestCaseData(
-                    new Uri("http://my.com/sizzle.js"),
-                    TimeSpan.FromSeconds(1),
-                    new object[] { false }).Throws(typeof(WebDriverTimeoutException));
-            }
-        }
-
         [TestCaseSource("LoadJQueryTestCases")]
         public void LoadJQuery(string version, TimeSpan? timeout, IEnumerable<object> mockValueSequence)
         {
@@ -83,24 +59,6 @@
             var sequence = mock.As<IJavaScriptExecutor>().SetupSequence(x => x.ExecuteScript(It.IsAny<string>()));
             mockValueSequence.Aggregate(sequence, (current, mockValue) => current.Returns(mockValue));
             mock.Object.JQuery().Load(jQueryUri, timeout);
-        }
-
-        [TestCaseSource("LoadSizzleTestCases")]
-        public void LoadSizzle(string version, TimeSpan? timeout, IEnumerable<object> mockValueSequence)
-        {
-            var mock = new Mock<IWebDriver>();
-            var sequence = mock.As<IJavaScriptExecutor>().SetupSequence(x => x.ExecuteScript(It.IsAny<string>()));
-            mockValueSequence.Aggregate(sequence, (current, mockValue) => current.Returns(mockValue));
-            mock.Object.Sizzle().Load(version, timeout);
-        }
-
-        [TestCaseSource("LoadSizzleWithUriTestCases")]
-        public void LoadSizzleWithUri(Uri sizzleUri, TimeSpan? timeout, IEnumerable<object> mockValueSequence)
-        {
-            var mock = new Mock<IWebDriver>();
-            var sequence = mock.As<IJavaScriptExecutor>().SetupSequence(x => x.ExecuteScript(It.IsAny<string>()));
-            mockValueSequence.Aggregate(sequence, (current, mockValue) => current.Returns(mockValue));
-            mock.Object.Sizzle().Load(sizzleUri, timeout);
         }
 
         [Test]
@@ -137,99 +95,6 @@
         }
 
         [Test]
-        public void FindElementWithSizzle()
-        {
-            var element = new Mock<IWebElement>();
-            element.Setup(x => x.TagName).Returns("div");
-            var list = new List<IWebElement> { element.Object };
-            var mock = MockWebDriver("return Sizzle('div');", new ReadOnlyCollection<IWebElement>(list));
-            var result = mock.Object.FindElement(By.SizzleSelector("div"));
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("div", result.TagName);
-        }
-
-        [Test]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void FindElementWithSizzleArgumentNull()
-        {
-            var mock = new Mock<IWebDriver>();
-            mock.Object.FindElement((SizzleSelector)null);
-        }
-
-        [Test]
-        [ExpectedException(typeof(NoSuchElementException))]
-        public void FindElementWithSizzleNoSuchElement()
-        {
-            var mock = MockWebDriver();
-
-            mock.Object.FindElement(By.SizzleSelector("div"));
-        }
-
-        [Test]
-        [ExpectedException(typeof(NoSuchElementException))]
-        public void FindElementWithSizzleNoSuchElementEmptyResult()
-        {
-            var mock = MockWebDriver("return Sizzle('div');", Enumerable.Empty<IWebElement>());
-
-            mock.Object.FindElement(By.SizzleSelector("div"));
-        }
-
-        [Test]
-        public void FindElementWithQuerySelector()
-        {
-            var element = new Mock<IWebElement>();
-            element.Setup(x => x.TagName).Returns("div");
-            var list = new List<IWebElement> { element.Object };
-            var mock = MockWebDriver("return document.querySelectorAll('div');", new ReadOnlyCollection<IWebElement>(list));
-            var result = mock.Object.FindElement(By.QuerySelector("div"));
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("div", result.TagName);
-        }
-
-        [Test]
-        public void FindElementWithNestedQuerySelector()
-        {
-            var element = new Mock<IWebElement>();
-            element.Setup(x => x.TagName).Returns("span");
-            var list = new List<IWebElement> { element.Object };
-            var mock = MockWebDriver(
-                "return document.querySelectorAll('div').length === 0 ? [] : document.querySelectorAll('div')[0].querySelectorAll('span');", 
-                new ReadOnlyCollection<IWebElement>(list));
-            var result = mock.Object.FindElement(By.QuerySelector("span", By.QuerySelector("div")));
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual("span", result.TagName);
-        }
-
-        [Test]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void FindElementWithQuerySelectorArgumentNull()
-        {
-            var mock = new Mock<IWebDriver>();
-            mock.Object.FindElement((QuerySelector)null);
-        }
-
-        [Test]
-        [ExpectedException(typeof(NoSuchElementException))]
-        public void FindElementWithQuerySelectorNoSuchElement()
-        {
-            var mock = MockWebDriver();
-
-            mock.Object.FindElement(By.QuerySelector("div"));
-        }
-
-        [Test]
-        [ExpectedException(typeof(NoSuchElementException))]
-        public void FindElementWithQuerySelectorNoSuchElementEmptyResult()
-        {
-            var mock = MockWebDriver("return document.querySelectorAll('div');", Enumerable.Empty<IWebElement>());
-            
-            mock.Object.FindElement(By.QuerySelector("div"));
-        }
-
-        [Test]
         public void FindElementsWithJQuery()
         {
             var element1 = new Mock<IWebElement>();
@@ -259,74 +124,6 @@
             var list = new List<object>();
             var mock = MockWebDriver("return jQuery('.test').get();", new ReadOnlyCollection<object>(list));
             var result = mock.Object.FindElements(By.JQuerySelector(".test"));
-
-            Assert.AreEqual(0, result.Count);
-        }
-
-        [Test]
-        public void FindElementsWithSizzle()
-        {
-            var element1 = new Mock<IWebElement>();
-            element1.Setup(x => x.TagName).Returns("div");
-            element1.Setup(x => x.GetAttribute("class")).Returns("test");
-
-            var element2 = new Mock<IWebElement>();
-            element2.Setup(x => x.TagName).Returns("span");
-            element2.Setup(x => x.GetAttribute("class")).Returns("test");
-
-            var list = new List<IWebElement> { element1.Object, element2.Object };
-            var mock = MockWebDriver("return Sizzle('.test');", new ReadOnlyCollection<IWebElement>(list));
-            var result = mock.Object.FindElements(By.SizzleSelector(".test"));
-
-            Assert.AreEqual(2, result.Count);
-
-            Assert.AreEqual("div", result[0].TagName);
-            Assert.AreEqual("test", result[0].GetAttribute("class"));
-
-            Assert.AreEqual("span", result[1].TagName);
-            Assert.AreEqual("test", result[1].GetAttribute("class"));
-        }
-
-        [Test]
-        public void FindElementsWithSizzleNotExists()
-        {
-            var list = new List<object>();
-            var mock = MockWebDriver("return Sizzle('.test');", new ReadOnlyCollection<object>(list));
-            var result = mock.Object.FindElements(By.SizzleSelector(".test"));
-
-            Assert.AreEqual(0, result.Count);
-        }
-
-        [Test]
-        public void FindElementsWithQuerySelector()
-        {
-            var element1 = new Mock<IWebElement>();
-            element1.Setup(x => x.TagName).Returns("div");
-            element1.Setup(x => x.GetAttribute("class")).Returns("test");
-
-            var element2 = new Mock<IWebElement>();
-            element2.Setup(x => x.TagName).Returns("span");
-            element2.Setup(x => x.GetAttribute("class")).Returns("test");
-
-            var list = new List<IWebElement> { element1.Object, element2.Object };
-            var mock = MockWebDriver("return document.querySelectorAll('.test');", new ReadOnlyCollection<IWebElement>(list));
-            var result = mock.Object.FindElements(By.QuerySelector(".test"));
-
-            Assert.AreEqual(2, result.Count);
-
-            Assert.AreEqual("div", result[0].TagName);
-            Assert.AreEqual("test", result[0].GetAttribute("class"));
-
-            Assert.AreEqual("span", result[1].TagName);
-            Assert.AreEqual("test", result[1].GetAttribute("class"));
-        }
-
-        [Test]
-        public void FindElementsWithQuerySelectorNotExists()
-        {
-            var list = new List<object>();
-            var mock = MockWebDriver("return document.querySelectorAll('.test');", new ReadOnlyCollection<object>(list));
-            var result = mock.Object.FindElements(By.QuerySelector(".test"));
 
             Assert.AreEqual(0, result.Count);
         }
@@ -673,32 +470,6 @@
         }
 
         [Test]
-        [ExpectedException(typeof(QuerySelectorNotSupportedException))]
-        public void QuerySelectorNotSupported()
-        {
-            var mock = new Mock<IWebDriver>();
-            mock.As<IJavaScriptExecutor>()
-                .Setup(x => x.ExecuteScript("return typeof document.querySelectorAll === 'function';")).Returns(false);
-            mock.Object.QuerySelector().CheckSupport();
-        }
-
-        [Test]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void CheckSelectorPrerequisitesWithoutLoader()
-        {
-            var mock = new Mock<IWebDriver>();
-            mock.Object.CheckSelectorPrerequisites(null);
-        }
-
-        [Test]
-        [ExpectedException(typeof(ArgumentNullException))]
-        public void LoadExternalLibraryWithoutLoader()
-        {
-            var mock = new Mock<IWebDriver>();
-            mock.Object.LoadExternalLibrary(null, null);
-        }
-
-        [Test]
         [ExpectedException(typeof(ArgumentNullException))]
         public void HelperWithNullDriver()
         {
@@ -715,10 +486,6 @@
 
             mock.As<IJavaScriptExecutor>().Setup(x => x.ExecuteScript("return typeof window.jQuery === 'function';"))
                 .Returns(true);
-            mock.As<IJavaScriptExecutor>().Setup(x => x.ExecuteScript("return typeof window.Sizzle === 'function';"))
-                .Returns(true);
-            mock.As<IJavaScriptExecutor>()
-                .Setup(x => x.ExecuteScript("return typeof document.querySelectorAll === 'function';")).Returns(true);
             return mock;
         }
     }
