+++ test/Selenium.WebDriver.Extensions.JQuery.Tests/Models/PositionTests.cs
@@ -3,6 +3,7 @@
     using NUnit.Framework;
     [TestFixture]
     [Category("Unit Tests")]
+
 #if !NET35
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
 #endif
