+++ test/Selenium.WebDriver.Extensions.JQuery.Tests/Selenium.WebDriver.Extensions.JQuery.Tests.csproj
@@ -107,6 +107,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Extensions\WebDriverExtensionsTests.cs" />
+    <Compile Include="Extensions\WebElementExtensionsTests.cs" />
     <Compile Include="Models\ByTests.cs" />
     <Compile Include="Loaders\JQueryLoaderTests.cs" />
     <Compile Include="Models\ChainJQueryHelperTests.cs" />
