+++ test/Selenium.WebDriver.Extensions.Sizzle.Tests/Selectors/SizzleSelectorTests.cs
@@ -3,6 +3,7 @@
     using System;
     using System.Collections;
     using NUnit.Framework;
+
     [TestFixture]
     [Category("Unit Tests")]
 #if !NET35
