+++ test/Selenium.WebDriver.Extensions.Sizzle.Tests/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Moq" version="4.2.1502.0911" targetFramework="net45" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
   <package id="Selenium.WebDriver" version="2.45.0" targetFramework="net45" />
 </packages>
