+++ test/Selenium.WebDriver.Extensions.Tests/Selenium.WebDriver.Extensions.Tests.csproj
@@ -106,11 +106,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Models\PositionTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Models\ByTests.cs" />
-    <Compile Include="Extensions\WebDriverExtensionsTests.cs" />
-    <Compile Include="Extensions\WebElementExtensionsTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
